Move CommandString nodes to their own file
